Added patrol point stuff

+++ Untitled Forest Game/Assets/Scripts/PatrolBehavior.cs
@@ -15,7 +15,7 @@ namespace Embers
             patrolPoints = new Transform[4];
             for (int i = 0; i < 4; i++)
             {
-                patrolPoints[i] = GameObject.FindGameObjectWithTag("World Generator").GetComponent<WorldGenerator>().patrolPoints[(int)enemyType, i];
+                patrolPoints[i] = GameObject.FindGameObjectWithTag("World Generator").GetComponent<WorldGenerator>().patrolPoints[enemyType, i];
             }
 
             SetDestinationToNextPatrolPoint();
